+++ src/features/comment/model/index.ts
@@ -1,2 +1,3 @@
 export * from "./commentQueryKey";
 export * from "./useCommentQuery";
+export * from "./useCommentStore";
