+++ src/features/comment/ui/index.ts
@@ -1,0 +1,4 @@
+export * from "./AddCommentDialog";
+export * from "./EditCommentDialog";
+export * from "./CommentList";
+export * from "./CommentRow";
