+++ src/features/index.ts
@@ -1,3 +1,4 @@
 export * from "./post";
 export * from "./user";
 export * from "./tag";
+export * from "./comment";
